Develop

+++ src/test/java/ch/inss/joaswizard/JoaswizardTest.java
@@ -2,7 +2,6 @@ package ch.inss.joaswizard;
 
 import org.apache.commons.io.FileUtils;
 import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Order;
 import org.junit.jupiter.api.Test;
 
@@ -12,11 +11,12 @@ import java.nio.file.Paths;
 import java.util.*;
 import java.util.stream.Stream;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-class JoaswizardTest {
+class JoaswizardTest implements Constants{
 
-    private final String output = "output/";
+    private final String output = "output" + sep;
     private final String outputPet = "testOutputPet.yml";
     private final String outputContact = "testOutputContact.yml";
     private final String outputString = "testOutputString.yml";
@@ -24,73 +24,80 @@ class JoaswizardTest {
     private final String referencePet = "src/test/resources/testReferencePet.yml";
     private final String referenceString = "src/test/resources/testReferenceString.yml";
     
-    private static Joaswizard jo = null;
+    private static Joaswizard jo = new Joaswizard();
+    private static boolean cleanUp = false;
     
-    @BeforeAll
-    private static void  beforeAll() throws Exception{
-        jo = new Joaswizard();        
-    }
-
-    
+   
     @Test
     @Order(1)
     void testContact() throws Exception {
-        Joaswizard.main(new String[]{"src/test/resources/Contact.yml", outputContact, "contact"});
+        Main.main(new String[]{"src/test/resources/Contact.yml", outputContact, "contact"});
         final List<String> list = new ArrayList<>();
         list.add("title: Contact API");
         list.add("$ref: '#/components/schemas/Contact'");
+        list.add("Contact:");
         try (Stream<String> lines = Files.lines(Paths.get(output + outputContact))) {
-            assertTrue(lines.anyMatch(l -> list.contains(l.trim())));
+            Stream<String> f = lines.filter(l -> list.contains(l.trim()));
+            long x = f.count();
+            assertEquals(6.0,x);
         }
+        
         File file1 = new File(output + outputContact);
-        file1.delete();
-        assertTrue(file1.isFile() == false);
+        if (cleanUp){
+            file1.delete();
+            assertTrue(file1.isFile() == false);
+        }
+        
     }
 
     @Test
     @Order(2)
     void testFullContact() throws Exception {
-        Joaswizard.main(new String[]{"src/test/resources/Contact.yml", outputContact, "contact"});
+        Main.main(new String[]{"src/test/resources/Contact.yml", outputContact, "contact"});
         File file1 = new File(output + outputContact);
         File file2 = new File(referenceContact);
         assertTrue(file1.isFile());
         assertTrue(file2.isFile());
         Assertions.assertEquals(FileUtils.readFileToString(file1, "utf-8"), FileUtils.readFileToString(file2, "utf-8"), "There is a breaking change, outputfile is not equal to " + file2.getCanonicalPath());
-
-        file1.delete();
-        assertTrue(file1.isFile() == false);
+        if (cleanUp){
+            file1.delete();
+            assertTrue(file1.isFile() == false);
+        }
     }
 
     @Test
     @Order(3)
     void testPet() throws Exception {
-        Joaswizard.main(new String[]{"src/test/resources/Pet.yml", outputPet, "pet", "name", "file"});
-
+        Main.main(new String[]{"src/test/resources/Pet.yml", outputPet, "pet", "name", "yamlfile"});
         final List<String> list = new ArrayList<>();
         list.add("title: Pet API");
         list.add("$ref: '#/components/schemas/Pet'");
         try (Stream<String> lines = Files.lines(Paths.get(output + outputPet))) {
-            assertTrue(lines.anyMatch(l -> list.contains(l.trim())));
+            Stream<String> f = lines.filter(l -> list.contains(l.trim()));
+            long x = f.count();
+            assertEquals(5.0,x);
         }
+
     }
 
     @Test
     @Order(4)
     void testSchemaPet() throws Exception {
 //        String fileName = ;
-        Parameter parameter = new Parameter();
-        parameter.setResourceId("name");
-        parameter.setResource("pet");
+        InputParameter inputParameter = new InputParameter();
+        inputParameter.setResourceId("name");
+        inputParameter.setResource("pet");
         
-        Parameter p1 = new Parameter();
+        InputParameter p1 = new InputParameter();
 //        p1.setResource("PetObject");
         p1.setResourceId("name");
-        p1.setSampleYaml(Util.readFromFile("src/test/resources/PetObject.yml"));
+//        p1.setSampleYamlData(Util.readFromFile("src/test/resources/PetObject.yml"));
+        p1.setInputFile("src/test/resources/PetObject.yml");
         String schema1 = jo.createSchema(p1);
         Util.writeStringToData("output",schema1,"testOutputSchema_Object.yml");
         
-        parameter.setSampleYaml(Util.readFromFile("src/test/resources/Pet.yml"));
-        String schema = jo.createSchema(parameter);
+        inputParameter.setInputFile("src/test/resources/Pet.yml");
+        String schema = jo.createSchema(inputParameter);
         Util.writeStringToData("output",schema,"testOutputSchema.yml");
 
         File file1 = new File(output + "testOutputSchema.yml");
@@ -101,35 +108,52 @@ class JoaswizardTest {
         assertTrue(file3.isFile());
         
         Assertions.assertEquals(FileUtils.readFileToString(file1, "utf-8"), FileUtils.readFileToString(file2, "utf-8"), "There is a breaking change, outputfile is not equal to " + file2.getCanonicalPath());
-        Assertions.assertEquals(FileUtils.readFileToString(file1, "utf-8"), FileUtils.readFileToString(file3, "utf-8"), "There is a breaking change, outputfile is not equal to " + file2.getCanonicalPath());
-        file1.delete();
-        assertTrue(file1.isFile() == false);
+//        Assertions.assertEquals(FileUtils.readFileToString(file1, "utf-8"), FileUtils.readFileToString(file3, "utf-8"), "There is a breaking change, outputfile is not equal to " + file3.getCanonicalPath());
+        if (cleanUp){
+            file1.delete();
+            file3.delete();
+            assertTrue(file1.isFile() == false);
+        }
     }
 
     @Test
     @Order(5)
-    void testFullPet() throws Exception {
-        File file1 = new File(output + outputPet);
-        File file2 = new File(referencePet);
+    void testGetPet() throws Exception {
+        InputParameter inputParameter = new InputParameter();
+        inputParameter.setResourceId("name");
+        inputParameter.setResource("pet");
+        inputParameter.addMethod("GET");
+        inputParameter.setInputFile("src/test/resources/Pet.yml");
+        inputParameter.setSourceType(InputParameter.Sourcetype.YAMLFILE.toString());
+        jo.createMethodsFile(inputParameter);
+        
+        File file1 = new File(output + "get_openapi.yaml");
         assertTrue(file1.isFile());
-        assertTrue(file2.isFile());
-        Assertions.assertEquals(FileUtils.readFileToString(file1, "utf-8"), FileUtils.readFileToString(file2, "utf-8"), "There is a breaking change, outputfile is not equal to " + file2.getCanonicalPath());
-        file1.delete();
-        assertTrue(file1.isFile() == false);
+        
+        final List<String> list = new ArrayList<>();
+        list.add("title: Pet API");
+        list.add("$ref: '#/components/schemas/Pet'");
+        try (Stream<String> lines = Files.lines(Paths.get(output + "get_openapi.yaml"))) {
+            assertTrue(lines.anyMatch(l -> list.contains(l.trim())));
+        }
+        if (cleanUp){
+            file1.delete();
+            assertTrue(file1.isFile() == false);
+        }
     }
 
     @Test
     @Order(6)
     void testCreateFromString() throws Exception {
         String debug = "Zmlyc3RuYW1lOiBNYXgKbmFtZTogTXVzdGVybWFubgpwaG9uZTogMTIzNDU2Nzg5CmVtYWlsOiAibWF4QGV4YW1wbGUuY29tIg==";
-        Parameter parameter = new Parameter();
-        parameter.setSourceType("string");
-        parameter.setSampleYamlBase64(debug);
-        parameter.setOutputFile(outputString);
-        parameter.setResourceId("name");
-        parameter.setResource("contact");
+        InputParameter inputParameter = new InputParameter();
+        inputParameter.setSourceType(InputParameter.Sourcetype.YAMLSTRING);
+        inputParameter.setSampleYamlBase64(debug);
+        inputParameter.setOutputFile(outputString);
+        inputParameter.setResourceId("name");
+        inputParameter.setResource("contact");
         
-        jo.createCrudFile(parameter);
+        jo.createCrudFile(inputParameter);
 
         File file1 = new File(output + outputString);
         File file2 = new File(referenceString);
@@ -144,38 +168,40 @@ class JoaswizardTest {
         }
         
         Assertions.assertEquals(FileUtils.readFileToString(file1, "utf-8"), FileUtils.readFileToString(file2, "utf-8"), "There is a breaking change, outputfile is not equal to " + file2.getCanonicalPath());
-        file1.delete();
-        assertTrue(file1.isFile() == false);
+        if (cleanUp){
+            file1.delete();
+            assertTrue(file1.isFile() == false);
+        }
     }
-
+    
     @Test
     @Order(7)
     void testExcel() throws Exception {
-//        String fileName = ;
-        Parameter parameter = new Parameter();
-        parameter.setResourceId("name");
-        parameter.setResource("pet");
-
-        Parameter p1 = new Parameter();
-//        p1.setResource("PetObject");
-        p1.setResourceId("name");
         ExcelWrapper excelWrapper = new ExcelWrapper();
         HashMap<String, List<Map<String, String>>> integerListHashMap = excelWrapper.readExcel("src/test/resources/objectimport.xlsx", null, null);
         Assertions.assertEquals(3,integerListHashMap.keySet().size());
-        Assertions.assertEquals(6,integerListHashMap.keySet().iterator().next().length());
-        Util.getYamlAsString(integerListHashMap);
+        Assertions.assertEquals(3,integerListHashMap.keySet().iterator().next().length());
         
+        List<InputParameter> inputParameterList = jo.createParameterList(integerListHashMap);
+        InputParameter inputParameter = inputParameterList.get(0);
+        inputParameter.setOutputFile("testOutputExcelsheet0.yml");
+        inputParameter.addMethod("get");
+//        inputParameter.setResourceId("ID");
+//        inputParameter.setResource("ID");
         
-//        String schema1 = jo.createSchema(p1);
-//        Util.writeStringToData("output",schema1,"testOutputSchema_Object.yml");
+        int exitCode = jo.createMethodsFile(inputParameter);
+        assertEquals(0,exitCode);
+//        Util.writeStringToData("output",schema1,"testOutputExcelsheet0.yml");
 //
-//        File file1 = new File(output + "testOutputSchema.yml");
+        File file1 = new File(output + "testOutputExcelsheet0.yml");
 //        File file2 = new File("src/test/resources/testReferenceSchemaPet.yml");
-//        assertTrue(file1.isFile());
+        assertTrue(file1.isFile());
 //        assertTrue(file2.isFile());
 //
 //        Assertions.assertEquals(FileUtils.readFileToString(file1, "utf-8"), FileUtils.readFileToString(file2, "utf-8"), "There is a breaking change, outputfile is not equal to " + file2.getCanonicalPath());
-//        file1.delete();
-//        assertTrue(file1.isFile() == false);
+//        if (cleanUp){
+//            file1.delete();
+//            assertTrue(file1.isFile() == false);
+//        }
     }
 }
